backend/pyproject.toml
[project]
name = "job-finder"
version = "0.1.0"
description = "Job Hunter application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "sqlmodel>=0.0.27",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.20",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "langgraph>=0.6.11",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langchain-ollama>=0.3.10",
    "python-jobspy>=1.1.25",
    "pandas>=2.2.0",
    "langchain-google-genai>=0.0.6",
    "playwright>=1.40.0",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
]
